+++ test/Peast/Syntax/ES2020/files/modules/ImportMeta/ImportMeta2.js
@@ -1,0 +1,1 @@
+var url = import.meta.url;